update nvim config install script

+++ nvim/install.fish
@@ -1,9 +1,12 @@
 #!/usr/bin/env fish
 
-# Backup old nvim config
-mv ~/.config/nvim{,.bak}
-mv ~/.local/share/nvim{,.bak}
-mv ~/.local/state/nvim{,.bak}
-mv ~/.cache/nvim{,.bak}
+if test -e "~/.config/nvim" and test -e "~/.config/nvim.bak"
+  echo "Cancel LazyVim nvim configuration installation, because there is already an backup folder for the config."
+  exit
+end
+
+if test -e "~/.config/nvim"
+  mv ~/.config/nvim{,.bak}
+end
 
 git clone https://github.com/LazyVim/starter ~/.config/nvim
